chore(slug): update slug stories to sb v7
