+++ packages/carbon-web-components/.storybook/main.ts
@@ -67,6 +67,8 @@ const stories = glob.sync(
     '../src/**/skip-to-content.stories.ts',
     '../src/**/slider.mdx',
     '../src/**/slider.stories.ts',
+    '../src/**/slug.mdx',
+    '../src/**/slug.stories.ts',
     '../src/**/tag.mdx',
     '../src/**/tag.stories.ts',
     '../src/**/tooltip.mdx',
