+++ packages/carbon-web-components/src/components/slug/slug.mdx
@@ -1,5 +1,8 @@
-import { Props, Description } from '@storybook/addon-docs/blocks';
+import { ArgsTable, Meta, Markdown } from '@storybook/addon-docs/blocks';
 import { cdnJs, cdnCss } from '../../globals/internal/storybook-cdn';
+import * as SlugStories from './slug.stories';
+
+<Meta of={SlugStories} />
 
 # Slug
 
@@ -21,8 +24,8 @@ Here's a quick example to get you started.
 import '@carbon/web-components/es/components/slug/index.js';
 ```
 
-<Description markdown={`${cdnJs({ components: ['slug'] })}`} />
-<Description markdown={`${cdnCss()}`} />
+<Markdown>{`${cdnJs({ components: ['slug'] })}`}</Markdown>
+<Markdown>{`${cdnCss()}`}</Markdown>
 
 ### HTML
 
@@ -49,4 +52,4 @@ Note: For `boolean` attributes, `true` means simply setting the attribute (e.g.
 `<cds-slug revert-active>`) and `false` means not setting the attribute (e.g.
 `<cds-slug>` without `revertActive` attribute).
 
-<Props of="cds-slug" />
+<ArgsTable of="cds-slug" />
